RDND_proper/home/dudy/Nehoray/RDND_proper/requirements.txt

Conversation opened. 1 read message.

Skip to content
Using GalEOAI Mail with screen readers
from:(eilonlok@galeoai.com)
Enable desktop notifications for GalEOAI Mail.
   OK  No thanks

12 of 19
requirments
Inbox

eilon lokhovsky
Attachments
Sun, Mar 12, 7:18 PM
to me

   
Translate message
Turn off for: Italian

 One attachment
  •  Scanned by Gmail
absl-py==1.0.0
albumentations==1.1.0
astropy==5.0.4
asttokens==2.1.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.0
brotlipy==0.7.0
cachetools==5.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cli-exit-tools==1.2.3
click==8.1.2
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
coloredlogs==15.0.1
colour-demosaicing==0.2.2
colour-science==0.4.1
cryptography==38.0.3
csaps==1.1.0
cycler==0.11.0
Cython==0.29.28
decorator==4.4.2
dill==0.3.4
dpcpp-cpp-rt==2022.2.1
easydict==1.9
einops==0.4.1
execnet==1.9.0
executing==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Fiona==1.9.1
flatbuffers==2.0
fonttools==4.32.0
future==0.18.2
gast==0.5.3
geopandas==0.12.2
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.6.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
humanfriendly==10.0
idna==3.4
image-registration==0.2.5
imageio==2.16.2
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.3
imutils==0.5.4
iniconfig==1.1.1
intel-cmplr-lib-rt==2022.2.1
intel-cmplr-lic-rt==2022.2.1
intel-opencl-rt==2022.2.1
intel-openmp==2022.2.1
ipython==8.6.0
jedi==0.18.1
joblib==1.2.0
keopscore==2.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
klepto==0.2.1
kornia==0.6.9
kornia-moons==0.2.0
lib-detect-testenv==2.0.2
lib-programname==2.0.4
libclang==13.0.0
lmdb==1.3.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.6
mkl==2022.2.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
moviepy==1.0.3
mpmath==1.2.1
munch==2.5.0
natsort==8.1.0
networkx==2.8
numpngw==0.1.0
numpy==1.22.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.8.0.121
nvidia-tensorrt==99.0.0
oauthlib==3.2.0
onnx==1.13.1
onnxruntime-gpu==1.14.1
opencv-python==4.7.0.72
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
paramiko==2.10.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pluggy==1.0.0
pox==0.3.0
proglog==0.1.10
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycocotools==2.0.6
pycparser==2.21
pyerfa==2.0.0.1
Pygments==2.13.0
pykeops==2.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.8
pyproj==3.4.1
PySocks==1.7.1
pytest==7.1.1
pytest-xdist==3.0.2
python-dateutil==2.8.2
pytube==12.1.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.1.3
scipy==1.8.0
scp==0.14.4
seaborn==0.12.1
sentry-sdk==1.16.0
setuptools==65.5.0
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.0
stack-data==0.6.0
sympy==1.11.1
tbb==2021.7.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
tensorrt==8.5.3.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.4.8
tomli==2.0.1
torch==1.13.0+cu116
torch-interpol==0.1.1
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
ultralytics==8.0.50
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.1.1
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.14.0
zipp==3.8.0
requirements.txt
Displaying requirements.txt.